[#83][fe][routine]refactor: 코드래빗 리뷰 반영
